fix galore
