+++ swift/trainers/optimizers/galore/utils.py
@@ -185,7 +185,7 @@ def get_optimizer(args: TrainingArguments, config: GaLoreConfig) -> Tuple[Any, A
         from .adafactor import GaLoreAdafactor
         optimizer_cls = GaLoreAdafactor
         optimizer_kwargs.update({'scale_parameter': False, 'relative_step': False})
-    elif args.optim in ('adamw_hf', 'adamw_torch'):
+    elif args.optim in ('adamw_hf', 'adamw_torch', 'adamw_torch_fused'):
         if config.quantize:
             assert importlib.util.find_spec('q_galore_torch') is not None, \
                 'Please install q-galore by `pip install q_galore_torch`'
