+++ swift/utils/__init__.py
@@ -11,9 +11,10 @@ from .np_utils import get_seed, stat_array, transform_jsonl_to_df
 from .tb_utils import TB_COLOR, TB_COLOR_SMOOTH, plot_images, read_tensorboard_file, tensorboard_smoothing
 from .torch_utils import (Serializer, activate_parameters, check_shared_disk, disable_safe_ddp_context_use_barrier,
                           empty_cache, find_all_linears, find_embedding, find_layers, find_norm, freeze_parameters,
-                          gc_collect, get_current_device, get_device, get_device_count, get_model_parameter_info,
-                          get_n_params_grads, init_process_group, safe_ddp_context, seed_worker, set_default_ddp_config,
-                          set_device, show_layers, time_synchronize, unwrap_model_for_generation)
+                          gc_collect, get_cu_seqlens_from_position_ids, get_current_device, get_device,
+                          get_device_count, get_model_parameter_info, get_n_params_grads, init_process_group,
+                          safe_ddp_context, seed_worker, set_default_ddp_config, set_device, show_layers,
+                          time_synchronize, unwrap_model_for_generation)
 from .utils import (add_version_to_work_dir, check_json_format, copy_files_by_pattern, deep_getattr, find_free_port,
                     format_time, get_env_args, import_external_file, json_parse_to_dict, lower_bound, parse_args,
                     patch_getattr, read_multi_line, remove_response, seed_everything, split_list, subprocess_run,
